e2e Move the stores into dedicated folders.
