+++ matrix-sdk/src/androidTest/java/org/matrix/androidsdk/CryptoTestHelper.java
@@ -19,8 +19,8 @@ package org.matrix.androidsdk;
 import android.content.Context;
 import android.net.Uri;
 
-import org.matrix.androidsdk.data.IMXStore;
-import org.matrix.androidsdk.data.MXFileStore;
+import org.matrix.androidsdk.data.store.IMXStore;
+import org.matrix.androidsdk.data.store.MXFileStore;
 import org.matrix.androidsdk.listeners.MXEventListener;
 import org.matrix.androidsdk.rest.callback.ApiCallback;
 import org.matrix.androidsdk.rest.client.LoginRestClient;
