+++ matrix-sdk/src/main/java/org/matrix/androidsdk/crypto/MXOlmDevice.java
@@ -26,8 +26,8 @@ import org.json.JSONObject;
 import org.matrix.androidsdk.crypto.algorithms.MXDecryptionResult;
 import org.matrix.androidsdk.crypto.data.MXKey;
 import org.matrix.androidsdk.crypto.data.MXOlmInboundGroupSession;
-import org.matrix.androidsdk.data.IMXCryptoStore;
-import org.matrix.androidsdk.data.IMXStore;
+import org.matrix.androidsdk.data.cryptostore.IMXCryptoStore;
+import org.matrix.androidsdk.data.store.IMXStore;
 import org.matrix.androidsdk.util.JsonUtils;
 import org.matrix.olm.OlmAccount;
 import org.matrix.olm.OlmManager;
