+++ matrix-sdk/src/main/java/org/matrix/androidsdk/data/DataRetriever.java
@@ -18,6 +18,7 @@ package org.matrix.androidsdk.data;
 import android.os.Looper;
 import android.text.TextUtils;
 
+import org.matrix.androidsdk.data.store.IMXStore;
 import org.matrix.androidsdk.rest.callback.ApiCallback;
 import org.matrix.androidsdk.rest.callback.SimpleApiCallback;
 import org.matrix.androidsdk.rest.client.RoomsRestClient;
