+++ matrix-sdk/src/main/java/org/matrix/androidsdk/data/store/MXFileStoreMetaData.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.matrix.androidsdk.data;
+package org.matrix.androidsdk.data.store;
 
 import org.matrix.androidsdk.rest.model.ThirdPartyIdentifier;
